+++ src/components/Home/LoginComponent.js
@@ -1,10 +1,13 @@
+import { fonts } from "../../styles/typography";
+
+
 export function LoginComponent(){
   
  return(
-    <div style={{marginTop: 50, marginRight:50}}>
-       <img src="/Image/loginsymbol.png" alt="loginImage"/>
-       <li>Login</li>
+    <div style={{ marginTop: 110, marginRight:50, textAlign: "right",}}>
+        <img src="/Images/loginsymbol.png" align="right" hspace="10px" width="4%"/>
+        <h3 style={{color: "white", letterSpacing: 1, }}>Login</h3>
     </div>
 
- );
+   );
 }
