+++ src/pages/Login/Login.js
@@ -1,9 +1,20 @@
-import { Wrapper } from "./styles";
+import { Wrapper, LoginContainer} from "../../components/Login/styles";
+import RegisterForm from "../../components/Login/RegisterForm";
 
 export default function Login(){
+ 
+
+
     return(
         <Wrapper>
-            Hola mundo
-        </Wrapper>
-    )
+
+            <LoginContainer>
+
+             <RegisterForm/>
+
+            </LoginContainer>
+      </Wrapper>
+
+    
+    );
 }
