--- src/pages/Login/styles.js
@@ -1,3 +1,0 @@
-import styled from "@emotion/styled";
-
-export const Wrapper = styled.div``
